fix: change correct file name + feat: update embed for the global match points
